refactor: extract a base class,and helper method for three class
