+++ src/Services/Masa.Dcc.Service/Domain/App/Aggregates/AppConfigObject.cs
@@ -4,24 +4,12 @@
 namespace Masa.Dcc.Service.Admin.Domain.App.Aggregates
 {
     [Table("AppConfigObjects")]
-    public class AppConfigObject : FullAggregateRoot<int, Guid>
+    public class AppConfigObject : GeneralConfigObject
     {
-        [Comment("EnvironmentClusterId")]
-        [Required]
-        [Range(1, int.MaxValue)]
-        public int EnvironmentClusterId { get; private set; }
-
         [Comment("AppId")]
         [Required]
         [Range(1, int.MaxValue)]
         public int AppId { get; private set; }
-
-        [Comment("ConfigObjectId")]
-        [Required]
-        [Range(1, int.MaxValue)]
-        public int ConfigObjectId { get; private set; }
-
-        public ConfigObject ConfigObject { get; set; } = null!;
 
         public AppConfigObject(int appId, int environmentClusterId)
         {
